implement edlib-based alignment of overlaps
This should greatly improve the scalability of the alignment, and allow
us to directly process extremely long overlaps without the memory issues
we ran into previously.

Fully exploiting this will require setting a bandwidth to the alignment,
but even without we should be able to run on overlaps of 50-100kb.